+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/IDBUtils.kt
@@ -579,6 +579,8 @@ interface IDBUtils {
 
     fun moveToGallery(context: Context, assetId: String, galleryId: String): AssetEntity
 
+    fun updateDateTaken(context: Context, assetId: String, timestamp: Int): Boolean
+
     fun getSomeInfo(context: Context, assetId: String): Pair<String, String?>?
 
     fun getUri(id: Long, type: Int, isOrigin: Boolean = false): Uri {